Rename Notification components
