+++ react/NotificationInline.js
@@ -1,3 +1,3 @@
-import NotificationInline from './components/Notification/notificationInline'
+import NotificationInline from './components/Notification/NotificationInline'
 
 export default NotificationInline
