+++ react/components/Notification/NotificationContent.tsx
